backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pymysql== 1.1.1
pydantic==2.7.2
python-dotenv
cryptography==42.0.7
PyJWT==2.8.0
itsdangerous==2.2.0